Add disguises
Place outfits on the level; they're generally in the dead-end rooms.
Bumping an outfit on the level will swap player outfits with it, if it differs from the player's current outfit.
One guard on the level may be a "seer" who can see through the disguise.
Add hotkey Ctrl+L for collecting all loot on the level.
Add hotkey Ctrl+D for swapping disguise mode.